move landingpage components to _components directory
