+++ src/app/(landing)/_components/feature-icons.tsx
