+++ src/app/(landing)/_components/footer.tsx
