+++ src/app/(landing)/_components/header.tsx
@@ -19,7 +19,7 @@ const routes = [
   },
 ] as const;
 
-const Header = () => {
+export const Header = () => {
   return (
     <header className="px-2 py-4 lg:py-6">
       <div className="container flex items-center gap-2 p-0">
@@ -69,5 +69,3 @@ const Header = () => {
     </header>
   );
 };
-
-export default Header;
