+++ src/app/(landing)/layout.tsx
@@ -1,8 +1,8 @@
 import { APP_TITLE } from "@/lib/constants";
 import { type ReactNode } from "react";
 import { type Metadata } from "next";
-import Header from "./header";
-import { Footer } from "./footer";
+import { Header } from "./_components/header";
+import { Footer } from "./_components/footer";
 
 export const metadata: Metadata = {
   title: APP_TITLE,
